restructured packages inside model to actually make sense
