+++ src/main/java/seedu/address/events/model/LocalModelChangedEvent.java
@@ -1,7 +1,7 @@
 package seedu.address.events.model;
 
 import seedu.address.events.BaseEvent;
-import seedu.address.model.datatypes.ReadOnlyAddressBook;
+import seedu.address.model.ReadOnlyAddressBook;
 
 /** Indicates data in the model has changed*/
 public class LocalModelChangedEvent extends BaseEvent {
