+++ src/main/java/seedu/address/model/ModelManager.java
@@ -5,12 +5,12 @@ import javafx.collections.transformation.FilteredList;
 import seedu.address.events.model.LocalModelChangedEvent;
 import seedu.address.exceptions.DuplicateTagException;
 import seedu.address.main.ComponentManager;
-import seedu.address.model.datatypes.AddressBook;
-import seedu.address.model.datatypes.ReadOnlyAddressBook;
 import seedu.address.model.person.Person;
 import seedu.address.model.person.ReadOnlyPerson;
 import seedu.address.model.person.UniquePersonList;
 import seedu.address.model.person.UniquePersonList.PersonNotFoundException;
+import seedu.address.model.tag.Tag;
+import seedu.address.model.tag.UniqueTagList;
 import seedu.address.parser.expr.Expr;
 import seedu.address.util.Config;
 import seedu.address.util.LoggerManager;
