+++ src/main/java/seedu/address/parser/Parser.java
@@ -3,8 +3,8 @@ package seedu.address.parser;
 import seedu.address.commands.*;
 import seedu.address.controller.Ui;
 import seedu.address.exceptions.IllegalValueException;
-import seedu.address.model.Tag;
-import seedu.address.model.UniqueTagList;
+import seedu.address.model.tag.Tag;
+import seedu.address.model.tag.UniqueTagList;
 import seedu.address.model.person.*;
 
 import java.util.Arrays;
