+++ src/test/java/seedu/address/commons/JsonUtilTest.java
@@ -1,16 +1,8 @@
 package seedu.address.commons;
 
 import org.junit.Test;
-import seedu.address.model.Tag;
-import seedu.address.model.datatypes.AddressBook;
-import seedu.address.model.datatypes.ReadOnlyAddressBook;
-import seedu.address.model.person.ReadOnlyPerson;
-import seedu.address.storage.StorageAddressBook;
 
 import java.io.IOException;
-import java.time.LocalDate;
-import java.util.ArrayList;
-import java.util.Arrays;
 
 import static org.junit.Assert.assertEquals;
