+++ src/test/java/seedu/address/commons/XmlUtilTest.java
@@ -3,7 +3,7 @@ package seedu.address.commons;
 import org.junit.Rule;
 import org.junit.Test;
 import org.junit.rules.ExpectedException;
-import seedu.address.model.datatypes.AddressBook;
+import seedu.address.model.AddressBook;
 import seedu.address.storage.StorageAddressBook;
 import seedu.address.testutil.AddressBookBuilder;
 import seedu.address.testutil.TestUtil;
