+++ src/test/java/seedu/address/storage/StorageManagerTest.java
@@ -12,8 +12,8 @@ import seedu.address.exceptions.DataConversionException;
 import seedu.address.exceptions.DuplicateTagException;
 import seedu.address.model.ModelManager;
 import seedu.address.model.UserPrefs;
-import seedu.address.model.datatypes.AddressBook;
-import seedu.address.model.datatypes.ReadOnlyAddressBook;
+import seedu.address.model.AddressBook;
+import seedu.address.model.ReadOnlyAddressBook;
 import seedu.address.model.person.UniquePersonList;
 import seedu.address.testutil.AddressBookBuilder;
 import seedu.address.testutil.SerializableTestClass;
