+++ src/test/java/seedu/address/testutil/PersonBuilder.java
@@ -1,7 +1,7 @@
 package seedu.address.testutil;
 
 import seedu.address.exceptions.IllegalValueException;
-import seedu.address.model.Tag;
+import seedu.address.model.tag.Tag;
 import seedu.address.model.person.*;
 
 /**
